Rename gazebo compatibility macros
