+++ src/barometer_plugin.cpp
@@ -15,7 +15,7 @@
  */
 
 #include "rosflight_plugins/barometer_plugin.h"
-#include "rosflight_plugins/gazebo_compat.h"
+#include "rosflight_plugins/gz_compat.h"
 
 namespace rosflight_plugins
 {
