+++ src/odometry_plugin.cpp
@@ -16,7 +16,7 @@
 
 
 #include "rosflight_plugins/odometry_plugin.h"
-#include "rosflight_plugins/gazebo_compat.h"
+#include "rosflight_plugins/gz_compat.h"
 
 namespace rosflight_plugins
 {
